+++ src/main/java/es/ujaen/dae/indicenciasurbanas/app/DaePracticaApplication.java
@@ -3,6 +3,7 @@ package es.ujaen.dae.indicenciasurbanas.app;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
@@ -10,6 +11,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @ComponentScan(basePackages = {"es.ujaen.dae.indicenciasurbanas"})
 @EntityScan(basePackages = {"es.ujaen.dae.indicenciasurbanas"})
 @EnableJpaRepositories(basePackages = "es.ujaen.dae.indicenciasurbanas")
+@EnableCaching
 public class DaePracticaApplication {
     public static void main(String[] args) {